# back/docs/swagger.yaml
basePath: /
definitions:
  enum.ModelPortDirection:
    enum:
    - in
    - out
    type: string
    x-enum-varnames:
    - ModelPortDirectionIn
    - ModelPortDirectionOut
  enum.ModelType:
    enum:
    - atomic
    - coupled
    type: string
    x-enum-varnames:
    - Atomic
    - Coupled
  json.ModelColors:
    properties:
      bodyBackgroundColor:
        type: string
      headerBackgroundColor:
        type: string
      headerTextColor:
        type: string
    type: object
  json.ModelComponent:
    properties:
      instanceId:
        type: string
      instanceMetadata:
        $ref: '#/definitions/json.ModelMetadata'
      modelId:
        type: string
    required:
    - instanceId
    - modelId
    type: object
  json.ModelConnection:
    properties:
      from:
        $ref: '#/definitions/json.ModelLink'
      to:
        $ref: '#/definitions/json.ModelLink'
    required:
    - from
    - to
    type: object
  json.ModelLink:
    properties:
      instanceId:
        type: string
      port:
        type: string
    required:
    - instanceId
    - port
    type: object
  json.ModelMetadata:
    properties:
      alwaysShowExtraInfo:
        type: boolean
      alwaysShowToolbar:
        type: boolean
      backgroundColor:
        type: string
      modelColors:
        $ref: '#/definitions/json.ModelColors'
      parameters:
        items:
          $ref: '#/definitions/json.ModelParameter'
        type: array
      position:
        $ref: '#/definitions/json.ModelPosition'
      style:
        $ref: '#/definitions/json.ModelStyle'
      toolbarPosition:
        $ref: '#/definitions/json.ToolbarPosition'
      toolbarVisible:
        type: boolean
    required:
    - position
    - style
    type: object
  json.ModelParameter:
    properties:
      description:
        type: string
      name:
        type: string
      type:
        $ref: '#/definitions/json.ParameterType'
      value: {}
    required:
    - name
    - type
    - value
    type: object
  json.ModelPort:
    properties:
      id:
        type: string
      type:
        $ref: '#/definitions/enum.ModelPortDirection'
    required:
    - id
    - type
    type: object
  json.ModelPosition:
    properties:
      x:
        type: number
      "y":
        type: number
    required:
    - x
    - "y"
    type: object
  json.ModelStyle:
    properties:
      height:
        type: number
      width:
        type: number
    required:
    - height
    - width
    type: object
  json.ParameterType:
    enum:
    - int
    - float
    - bool
    - string
    - object
    type: string
    x-enum-varnames:
    - ParameterTypeInt
    - ParameterTypeFloat
    - ParameterTypeBool
    - ParameterTypeString
    - ParameterTypeObject
  json.ToolbarPosition:
    enum:
    - top
    - left
    - right
    - bottom
    type: string
    x-enum-varnames:
    - ToolbarPositionTop
    - ToolbarPositionLeft
    - ToolbarPositionRight
    - ToolbarPositionBottom
  model.Diagram:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      modelId:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      workspaceId:
        type: string
    required:
    - name
    type: object
  model.Library:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      models:
        items:
          $ref: '#/definitions/model.Model'
        type: array
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  model.Model:
    properties:
      code:
        type: string
      components:
        items:
          $ref: '#/definitions/json.ModelComponent'
        type: array
      connections:
        items:
          $ref: '#/definitions/json.ModelConnection'
        type: array
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      libId:
        type: string
      metadata:
        $ref: '#/definitions/json.ModelMetadata'
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/json.ModelPort'
        type: array
      type:
        $ref: '#/definitions/enum.ModelType'
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  model.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      fullname:
        type: string
      id:
        type: string
      libraries:
        items:
          $ref: '#/definitions/model.Library'
        type: array
      modelTypes:
        items:
          $ref: '#/definitions/model.Model'
        type: array
      password:
        maxLength: 50
        minLength: 6
        type: string
      refresh_token:
        type: string
      updatedAt:
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
      workspaces:
        items:
          $ref: '#/definitions/model.Workspace'
        type: array
    required:
    - email
    - password
    - username
    type: object
  model.Workspace:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      diagrams:
        items:
          $ref: '#/definitions/model.Diagram'
        type: array
      id:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  request.ModelRequest:
    properties:
      code:
        type: string
      components:
        items:
          $ref: '#/definitions/json.ModelComponent'
        type: array
      connections:
        items:
          $ref: '#/definitions/json.ModelConnection'
        type: array
      description:
        type: string
      id:
        type: string
      libId:
        type: string
      metadata:
        $ref: '#/definitions/json.ModelMetadata'
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/json.ModelPort'
        type: array
      type:
        $ref: '#/definitions/enum.ModelType'
    required:
    - code
    - components
    - connections
    - description
    - metadata
    - name
    - ports
    - type
    type: object
  request.PasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
...
  request.UpdateUserRequest:
    properties:
      names:
        type: string
    required:
    - names
    type: object
...
  response.ModelResponse:
    properties:
      code:
        type: string
      components:
        items:
          $ref: '#/definitions/json.ModelComponent'
        type: array
      connections:
        items:
          $ref: '#/definitions/json.ModelConnection'
        type: array
      description:
        type: string
      id:
        type: string
      libId:
        type: string
      metadata:
        $ref: '#/definitions/json.ModelMetadata'
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/json.ModelPort'
        type: array
      type:
        $ref: '#/definitions/enum.ModelType'
      userId:
        type: string
    required:
    - code
    - components
    - connections
    - description
    - id
    - metadata
    - name
    - ports
    - type
    - userId
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is the API documentation for Easy DEVS.
  title: Easy DEVS API
  version: "1.0"
paths:
...
  /model:
    get:
      description: Retrieve a list of all models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ModelResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all models
      tags:
      - models
    post:
      consumes:
      - application/json
      description: Create a new model entry
      parameters:
      - description: Model data
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/request.ModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ModelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a model
      tags:
      - models
  /model/{id}:
    delete:
      description: Delete a model by its ID
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Delete a model
      tags:
      - models
    get:
      description: Retrieve a single model by its ID
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ModelResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a model by ID
      tags:
      - models
    patch:
      consumes:
      - application/json
      description: Update an existing model with partial data
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/request.ModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ModelResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update a model
      tags:
      - models
  /model/{id}/recursive:
    get:
      description: Retrieve a single model by its ID
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.ModelResponse'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a model by ID
      tags:
      - models
  /model/{id}/simulate:
    get:
      description: generateSimulationFile generate a zip that will contain all infromations
        for simulation
      parameters:
      - description: Model ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Generate simulations files
      tags:
      - models
  /user:
    get:
      description: Retrieve a list of all user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all user
      tags:
      - user
  /user/{id}:
    delete:
      description: Delete an existing user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User password confirmation
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.PasswordRequest'
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a user by ID
      tags:
      - user
    get:
      description: Retrieve a single user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a user by ID
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Update an existing user with partial data
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Partial user update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a user
      tags:
      - user
...
swagger: "2.0"